% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjce_ler_cjpg.R
\name{tjce_ler_cjpg}
\alias{tjce_ler_cjpg}
\title{Lê decisões de primeira instância do TJCE baixadas por meio da função baixar_cjpg.}
\usage{
tjce_ler_cjpg(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Se NULL, informar diretorio}

\item{diretorio}{objeto ou diretório onde se encontram as páginas.}
}
\value{
uma tibble com nove colunas: processo, classe, assunto, magistrado,comarca, foro, vara,
disponibilizacao e julgado (texto da decisão).
}
\description{
Lê decisões de primeira instância do TJCE baixadas por meio da função baixar_cjpg.
}
\examples{
\dontrun{
ler_cjpg()
}
}
