% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjce_ler_tabela_docs.R
\name{tjce_ler_tabela_docs}
\alias{tjce_ler_tabela_docs}
\title{Ler tabelas baixadas com tjce_baixar_tabela_docs}
\usage{
tjce_ler_tabela_docs(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{diretorio}{Informar se não informar diretório}
}
\value{
tibble
}
\description{
Ler tabelas baixadas com tjce_baixar_tabela_docs
}
